+++ CHANGELOG.md
@@ -23,6 +23,8 @@
   - `.JsClick(centerXOffset=0, centerYOffset=0)`
     - proper tests coverage is yet needed
 - made Configuration ThreadLocal
+- added SeleneElement methods:
+  - Matching(Condition)
 
 ## 1.0.0-alpha03 (to be released on 2020.06.03)
 - added `SeleneElement#Type(string keys)`, i.e. `S(selector).Type(text)`